fix: correlated-subquery column resolution, drop broken migrations, add idempotent deploy SQL
- admin dashboard / seasons list: qualify subquery refs as `seasons.id`
  (Drizzle was emitting unqualified `"id"`, which Postgres bound to the
  inner table's `id` column on `games`, making gameCount/completedGameCount
  always 0)
- cast COUNT(*) and COUNT(DISTINCT ...) to ::int so values arrive as numbers
- delete drizzle/0000_*..0002_* — the auto-generated full-schema snapshot
  cannot be applied to prod (tables already exist) and 0001 references a
  column never added in 0000
- scripts/deploy-schema.sql — idempotent prod migration: adds new columns
  to seasons/games/player_seasons, harmonizes _key→_unique constraint
  names, ensures `tbd` team row, backfills status + default_location

+++ app/api/bash/admin/seasons/[id]/route.ts
@@ -34,9 +34,9 @@ export async function GET(_request: NextRequest, context: RouteContext) {
 
   const [counts] = await db
     .select({
-      gameCount: sql<number>`(SELECT COUNT(*) FROM games WHERE season_id = ${id})`,
-      completedGameCount: sql<number>`(SELECT COUNT(*) FROM games WHERE season_id = ${id} AND status = 'final')`,
-      playerCount: sql<number>`(SELECT COUNT(DISTINCT player_id) FROM player_seasons WHERE season_id = ${id})`,
+      gameCount: sql<number>`(SELECT COUNT(*)::int FROM games WHERE season_id = ${id})`,
+      completedGameCount: sql<number>`(SELECT COUNT(*)::int FROM games WHERE season_id = ${id} AND status = 'final')`,
+      playerCount: sql<number>`(SELECT COUNT(DISTINCT player_id)::int FROM player_seasons WHERE season_id = ${id})`,
     })
     .from(sql`(SELECT 1) AS _`)
